+++ Pangle/src/main/java/com/mopub/mobileads/PangleAdapterConfiguration.java
@@ -170,16 +170,19 @@ public class PangleAdapterConfiguration extends BaseAdapterConfiguration {
 
     private static JSONArray getAdCallSource() {
         JSONArray adCallSource = new JSONArray();
+
         try {
             JSONObject mediationObject = new JSONObject();
-            mediationObject.putOpt("name","mediation");
-            mediationObject.putOpt("value","mopub");
+            mediationObject.putOpt("name", "mediation");
+            mediationObject.putOpt("value", "mopub");
             adCallSource.put(mediationObject);
+
             JSONObject adapterVersionObject = new JSONObject();
-            adapterVersionObject.putOpt("name","adapter_version");
-            adapterVersionObject.putOpt("value","1.2.0");
+            adapterVersionObject.putOpt("name", "adapter_version");
+            adapterVersionObject.putOpt("value", "1.2.0");
             adCallSource.put(adapterVersionObject);
-        } catch (Throwable ignore) {
+        } catch (Throwable exception) {
+            MoPubLog.log(CUSTOM, ADAPTER_NAME, "AdCallSource encounter parsing error: " + exception.getLocalizedMessage());
         }
         return adCallSource;
     }