+++ Pangle/CHANGELOG.md
@@ -1,7 +1,6 @@
 ## Changelog
   * 3.4.1.1.1
-     * This version of the adapters has been certified with Pangle 3.4.1.1 and MoPub SDK 5.15.0.
-     * Add ad sources
+     * Pass `AdCallSource` data for Pangle SDK to collect ad source information.
 
   * 3.4.1.1.0
      * This version of the adapters has been certified with Pangle 3.4.1.1 and MoPub SDK 5.15.0.